// dashboard/src/utils/hasura-api/generated/schemas/trackTableArgs.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing database operations, migrations, and metadata permissions
 * OpenAPI spec version: 1.0.0
 */
import type { TrackTableArgsTable } from './trackTableArgsTable';

export interface TrackTableArgs {
  table: TrackTableArgsTable;
  /** The data source name */
  source: string;
}