// dashboard/src/utils/hasura-api/generated/schemas/trackTableBulkOperation.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { TrackTableBulkOperationType } from './trackTableBulkOperationType';
import type { TrackTableStep } from './trackTableStep';

export interface TrackTableBulkOperation {
  type: TrackTableBulkOperationType;
  source?: string;
  version?: number;
  args: TrackTableStep[];
}