// dashboard/src/utils/hasura-api/generated/schemas/trackTableStep.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing database operations, migrations, and metadata permissions
 * OpenAPI spec version: 1.0.0
 */
import type { TrackTableStepType } from './trackTableStepType';
import type { TrackTableArgs } from './trackTableArgs';

export interface TrackTableStep {
  /** Use after creating a table with run_sql to track it. You can provide this in the same query as the run_sql step */
  type: TrackTableStepType;
  args: TrackTableArgs;
}