// dashboard/src/utils/hasura-api/generated/schemas/untrackTableStep.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { UntrackTableArgs } from './untrackTableArgs';
import type { UntrackTableStepType } from './untrackTableStepType';

export interface UntrackTableStep {
  type: UntrackTableStepType;
  args: UntrackTableArgs;
}